Move Singer IO to _singerlib
